+++ packages/transaction-controller/src/utils/batch.ts
@@ -1,7 +1,8 @@
 import type EthQuery from '@metamask/eth-query';
 import { rpcErrors } from '@metamask/rpc-errors';
 import type { Hex } from '@metamask/utils';
-import { createModuleLogger } from '@metamask/utils';
+import { bytesToHex, createModuleLogger } from '@metamask/utils';
+import { parse, v4 } from 'uuid';
 
 import {
   doesChainSupportEIP7702,
@@ -9,15 +10,24 @@ import {
   isAccountUpgradedToEIP7702,
 } from './eip7702';
 import {
+  getBatchSizeLimit,
   getEIP7702SupportedChains,
   getEIP7702UpgradeContractAddress,
 } from './feature-flags';
 import { validateBatchRequest } from './validation';
-import type { TransactionController, TransactionControllerMessenger } from '..';
+import {
+  determineTransactionType,
+  type TransactionBatchRequest,
+  type TransactionController,
+  type TransactionControllerMessenger,
+} from '..';
 import { projectLogger } from '../logger';
+import type {
+  NestedTransactionMetadata,
+  TransactionBatchSingleRequest,
+} from '../types';
 import {
   TransactionEnvelopeType,
-  type TransactionBatchRequest,
   type TransactionBatchResult,
   type TransactionParams,
   TransactionType,
@@ -29,6 +39,7 @@ type AddTransactionBatchRequest = {
   getEthQuery: (networkClientId: string) => EthQuery;
   getInternalAccounts: () => Hex[];
   messenger: TransactionControllerMessenger;
+  publicKeyEIP7702?: Hex;
   request: TransactionBatchRequest;
 };
 
@@ -36,6 +47,7 @@ type IsAtomicBatchSupportedRequest = {
   address: Hex;
   getEthQuery: (chainId: Hex) => EthQuery;
   messenger: TransactionControllerMessenger;
+  publicKeyEIP7702?: Hex;
 };
 
 const log = createModuleLogger(projectLogger, 'batch');
@@ -54,15 +66,25 @@ export async function addTransactionBatch(
     getChainId,
     getInternalAccounts,
     messenger,
+    publicKeyEIP7702,
     request: userRequest,
   } = request;
+
+  const sizeLimit = getBatchSizeLimit(messenger);
 
   validateBatchRequest({
     internalAccounts: getInternalAccounts(),
     request: userRequest,
+    sizeLimit,
   });
 
-  const { from, networkClientId, requireApproval, transactions } = userRequest;
+  const {
+    batchId: batchIdOverride,
+    from,
+    networkClientId,
+    requireApproval,
+    transactions,
+  } = userRequest;
 
   log('Adding', userRequest);
 
@@ -75,9 +97,14 @@ export async function addTransactionBatch(
     throw rpcErrors.internal('Chain does not support EIP-7702');
   }
 
+  if (!publicKeyEIP7702) {
+    throw rpcErrors.internal('EIP-7702 public key not specified');
+  }
+
   const { delegationAddress, isSupported } = await isAccountUpgradedToEIP7702(
     from,
     chainId,
+    publicKeyEIP7702,
     messenger,
     ethQuery,
   );
@@ -89,7 +116,12 @@ export async function addTransactionBatch(
     throw rpcErrors.internal('Account upgraded to unsupported contract');
   }
 
-  const nestedTransactions = transactions.map((tx) => tx.params);
+  const nestedTransactions = await Promise.all(
+    transactions.map((tx) =>
+      getNestedTransactionMeta(userRequest, tx, ethQuery),
+    ),
+  );
+
   const batchParams = generateEIP7702BatchTransaction(from, nestedTransactions);
 
   const txParams: TransactionParams = {
@@ -101,6 +133,7 @@ export async function addTransactionBatch(
     const upgradeContractAddress = getEIP7702UpgradeContractAddress(
       chainId,
       messenger,
+      publicKeyEIP7702,
     );
 
     if (!upgradeContractAddress) {
@@ -113,14 +146,15 @@ export async function addTransactionBatch(
 
   log('Adding batch transaction', txParams, networkClientId);
 
-  const { transactionMeta, result } = await addTransaction(txParams, {
+  const batchId = batchIdOverride ?? generateBatchId();
+
+  const { result } = await addTransaction(txParams, {
+    batchId,
     nestedTransactions,
     networkClientId,
     requireApproval,
     type: TransactionType.batch,
   });
-
-  const batchId = transactionMeta.id;
 
   // Wait for the transaction to be published.
   await result;
@@ -139,7 +173,16 @@ export async function addTransactionBatch(
 export async function isAtomicBatchSupported(
   request: IsAtomicBatchSupportedRequest,
 ): Promise<Hex[]> {
-  const { address, getEthQuery, messenger } = request;
+  const {
+    address,
+    getEthQuery,
+    messenger,
+    publicKeyEIP7702: publicKey,
+  } = request;
+
+  if (!publicKey) {
+    throw rpcErrors.internal('EIP-7702 public key not specified');
+  }
 
   const chainIds7702 = getEIP7702SupportedChains(messenger);
   const chainIds: Hex[] = [];
@@ -150,6 +193,7 @@ export async function isAtomicBatchSupported(
     const { isSupported, delegationAddress } = await isAccountUpgradedToEIP7702(
       address,
       chainId,
+      publicKey,
       messenger,
       ethQuery,
     );
@@ -162,4 +206,42 @@ export async function isAtomicBatchSupported(
   log('Atomic batch supported chains', chainIds);
 
   return chainIds;
+}
+
+/**
+ * Generate a tranasction batch ID.
+ *
+ * @returns  A unique batch ID as a hexadecimal string.
+ */
+function generateBatchId(): Hex {
+  const idString = v4();
+  const idBytes = new Uint8Array(parse(idString));
+  return bytesToHex(idBytes);
+}
+
+/**
+ * Generate the metadata for a nested transaction.
+ *
+ * @param request - The batch request.
+ * @param singleRequest - The request for a single transaction.
+ * @param ethQuery - The EthQuery instance used to interact with the Ethereum blockchain.
+ * @returns The metadata for the nested transaction.
+ */
+async function getNestedTransactionMeta(
+  request: TransactionBatchRequest,
+  singleRequest: TransactionBatchSingleRequest,
+  ethQuery: EthQuery,
+): Promise<NestedTransactionMetadata> {
+  const { from } = request;
+  const { params } = singleRequest;
+
+  const { type } = await determineTransactionType(
+    { from, ...params },
+    ethQuery,
+  );
+
+  return {
+    ...params,
+    type,
+  };
 }